+++ kamehouse-shell/bin/common/sudoers/www-data/exec-script.sh
@@ -20,7 +20,6 @@ if [ "$?" != "0" ]; then
   exit 99
 fi
 
-LOG_SCRIPT_RUN_TIME_IN_DEBUG=true
 LOG_PROCESS_TO_FILE=false
 SCRIPT=""
 SCRIPT_ARGS=""