+++ changelog.md
@@ -7,6 +7,7 @@
 
 #### v9.02
 
+- Updated log script runtime function
 - Added more bash colors
 - Added script name to log entries
 - Moved shell scripts to default to log output to file